chore: address CodeRabbit review findings
- transformComposer.ts / buildDragTransform: add @example tags to the
  public JSDoc per the repo's Google-style requirement.
- e2e: consolidate the tripled nextFrame/readRotation/sampleFrames/
  beginHorizontalDrag helpers into e2e/_helpers/transform.ts, which
  exists for exactly this cross-spec sharing.
- test pages: add @component to the three new characterization pages'
  script JSDoc, matching the repo's existing convention (script-level
  @component, not the HTML-comment form).
- drag.spec.ts: remove the appended element after the bound-MotionValue
  test so the DOM is restored for subsequent tests.
- guard report: the maintainer sign-off / PR gate is resolved (PR #446);
  the residual-risk entry said it was still outstanding.

Skipped two findings deliberately: SHA-pinning the new workflow actions
(the file's pre-existing steps use version tags with explicit zizmor
suppressions — changing only the new steps would be inconsistent; pin
repo-wide or not at all) and extracting shared CSS from the two docs
demos (docs demos are published as self-contained code samples by
convention).
