+++ e2e/motion/hover-authored-transforms.spec.ts
@@ -1,24 +1,7 @@
-import { expect, test, type Locator, type Page } from '@playwright/test'
+import { expect, test } from '@playwright/test'
+import { readRotation, sampleFrames } from '../_helpers/transform'
 
 const URL = '/tests/motion/hover-authored-transforms?@isPlaywright=true'
-
-const nextFrame = (page: Page) =>
-    page.evaluate(() => new Promise<void>((resolve) => requestAnimationFrame(() => resolve())))
-
-const readRotation = (card: Locator) =>
-    card.evaluate((element) => {
-        const matrix = new DOMMatrixReadOnly(getComputedStyle(element).transform)
-        return (Math.atan2(matrix.b, matrix.a) * 180) / Math.PI
-    })
-
-const sampleRotations = async (page: Page, card: Locator, count: number) => {
-    const samples: number[] = []
-    for (let frame = 0; frame < count; frame++) {
-        await nextFrame(page)
-        samples.push(await readRotation(card))
-    }
-    return samples
-}
 
 test.describe('motion/whileHover over authored transforms', () => {
     test.beforeEach(async ({ page }) => {
@@ -50,7 +33,7 @@ test.describe('motion/whileHover over authored transforms', () => {
         // Leave: move the pointer well away from the card.
         await page.mouse.move(4, 4)
 
-        const samples = await sampleRotations(page, card, 45)
+        const samples = await sampleFrames(page, () => readRotation(card), 45)
 
         // The restore must be continuous: settling to neutral and then
         // snapping to the authored angle shows up as a single-frame jump.
