+++ e2e/motion/pan-authored-transforms.spec.ts
@@ -1,37 +1,7 @@
-import { expect, test, type Locator, type Page } from '@playwright/test'
+import { expect, test } from '@playwright/test'
+import { beginHorizontalDrag as beginPan, readRotation, sampleFrames } from '../_helpers/transform'
 
 const URL = '/tests/motion/pan-authored-transforms?@isPlaywright=true'
-
-const nextFrame = (page: Page) =>
-    page.evaluate(() => new Promise<void>((resolve) => requestAnimationFrame(() => resolve())))
-
-const readRotation = (card: Locator) =>
-    card.evaluate((element) => {
-        const matrix = new DOMMatrixReadOnly(getComputedStyle(element).transform)
-        return (Math.atan2(matrix.b, matrix.a) * 180) / Math.PI
-    })
-
-const sampleRotations = async (page: Page, card: Locator, count: number) => {
-    const samples: number[] = []
-    for (let frame = 0; frame < count; frame++) {
-        await nextFrame(page)
-        samples.push(await readRotation(card))
-    }
-    return samples
-}
-
-const beginPan = async (page: Page, card: Locator) => {
-    await card.waitFor({ state: 'visible' })
-    await card.scrollIntoViewIfNeeded()
-    const box = await card.boundingBox()
-    if (!box) throw new Error('missing pan-card bounds')
-
-    const x = box.x + box.width / 2
-    const y = box.y + box.height / 2
-    await page.mouse.move(x, y)
-    await page.mouse.down()
-    await page.mouse.move(x + 56, y, { steps: 6 })
-}
 
 test.describe('motion/whilePan over authored transforms', () => {
     test.beforeEach(async ({ page }) => {
@@ -65,7 +35,7 @@ test.describe('motion/whilePan over authored transforms', () => {
         await page.waitForTimeout(600)
         await page.mouse.up()
 
-        const samples = await sampleRotations(page, card, 45)
+        const samples = await sampleFrames(page, () => readRotation(card), 45)
 
         // The restore must be continuous: settling to neutral and then
         // snapping to the authored angle shows up as a single-frame jump.
