+++ src/lib/utils/transformComposer.ts
@@ -17,6 +17,9 @@ const getLastValue = (values: readonly unknown[]): unknown => values.at(-1)
  * @param baseTransform An authored raw CSS transform that cannot be represented as channels.
  * @param transformTemplate Optional user transform template applied to the live values.
  * @returns The composed CSS transform string.
+ * @example
+ * buildGestureTransform({ x: '20px', rotate: 4 }, 'perspective(600px)')
+ * // => 'translateX(20px) rotate(4deg) perspective(600px)'
  */
 export const buildGestureTransform = (
     latestValues: GestureTransformValues,
@@ -32,6 +35,9 @@ export const buildGestureTransform = (
  *
  * @param source Motion target whose transform channels should be collected.
  * @returns Only recognized transform channels with scalar final values.
+ * @example
+ * collectGestureTransformValues({ rotate: [0, 8], scale: 1.1, opacity: 0.5 })
+ * // => { rotate: 8, scale: 1.1 } — keyframe arrays resolve to their final value
  */
 export const collectGestureTransformValues = (
     source: Record<string, unknown> | undefined
@@ -54,6 +60,9 @@ export const collectGestureTransformValues = (
  *
  * @param source Motion target to partition.
  * @returns Separate transform and native style records.
+ * @example
+ * splitGestureTransformValues({ rotate: 4, cursor: 'grabbing' })
+ * // => { transform: { rotate: 4 }, native: { cursor: 'grabbing' } }
  */
 export const splitGestureTransformValues = (source: Record<string, unknown>) => {
     const transform: Record<string, unknown> = {}
